Update PR comments
- moved all tests into one file
- applied suggestion to types to lessen duplication
